Messy commit to stash in-progress changes
Have been running a lot more tests, and am still struggling to get the MAE for fantasy points consistently below 5.9 with novel data, even after selecting which data to focus on using a correlation matrix as well as experimenting with various different models (primarily, from sklearn). This may be the feasible limit for this task, however, when graphing the findings, all models settle into generally predicting between a certain range, and do not even attempt to cover outliers ("breakout" fantasy point games, primarily, are what I want the model to be able to capture). Maybe there simply is no good way to predict it with the limited data points I am using. I will try to add more data and run it through the same correlation matrix to see if there are any other data points which may help.