Going to keep how it is
This model is too simple to truly be able to predict breakout or poor games. A full refactoring using more object orientation and/or better segmented program files would also be beneficial to continue scaling this up. I plan on applying the changes made to the quarterback model to the other position groups, then merging into main once that is done.

Overall, though, I consider an improvement to MAE of around 0.8 (from 6.5 to 5.7) to still be good results from simply incorporating more data. There also may be more varied (good or bad) results from the other positions, as well.